Added Queue system. Rate limiting
Everything is at a level i think im happy with.
the only thing that i think might be cool, but ultimately pointless
would have a way to split the messages over multiple api-keys.
i think if i were to scale this to servers that i dont control it would
be very useful. but as is, im happy